// packages/renderer/src/api/generated/types.gen.ts
export type Importance = 'none' | 'low' | 'medium' | 'high';
// ...
export type TagModel = {
    id: string;
    name: string;
    color?: string | null;
    user_id: string;
};

/**
 * TodoCreate
 */
export type TodoCreate = {
    item: string;
    description?: string | null;
    alarmtime?: string | null;
    starttime: string;
    endtime: string;
    importance?: Importance;
    tags?: Array<string> | null;
};

/**
 * TodoModel
 */
export type TodoModel = {
    id: string;
    item: string;
    description?: string | null;
    created_time: string;
    alarm_time?: string | null;
    start_time: string;
    end_time: string;
    importance: Importance;
    user_id: string;
    tags?: Array<string> | null;
};